% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGMAP_source.R
\name{data_simu}
\alias{data_simu}
\title{generate simulated hic_mat and true tads}
\usage{
data_simu(stype = "poisson-dist", nratio = 2.5, mu0 = 20, resl = 1)
}
\arguments{
\item{stype}{One of four types of simulated data in the manuscript:
poission-dist, poission-dist-hier, nb-dist, nb-dist-hier;
poission- or nb- indicates poission distribution or negative bionomial distribution
-hier indicated subtads are generated nestly}

\item{nratio}{The effect size between intra- and inter domain, larger means higher intra-tad contacts}

\item{mu0}{The mean parameter, default 20}

\item{resl}{Resolution, default set to 1}
}
\value{
A list includes following elements:
\item{hic_mat}{n by n contact matrix}
\item{hierTads}{True heirarchical domains}
\item{tads_true}{True TADs}
}
\description{
generate simulated hic_mat and true tads
}
