% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hic_rao_IMR90_chr15}
\alias{hic_rao_IMR90_chr15}
\title{Normalized Hi-C data for IMR90, chr15 with resolution 10kb.}
\format{A data table with 3 variables:
\describe{
  \item{n1}{bin 1}
  \item{n2}{bin 2}
  \item{count}{normalized counts}
}}
\source{
Rao et al., Cell 2014, A 3D map of the human genome at kilobase resolution reveals principles of chromatin looping
}
\usage{
hic_rao_IMR90_chr15
}
\description{
Normalized Hi-C data for IMR90, chr15 with resolution 10kb.
}
\keyword{datasets}
