% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGMAP_source.R
\name{plotdom}
\alias{plotdom}
\title{visualize hierarchical domains}
\usage{
plotdom(hic_dat, hiertads_gmap, start_bin, end_bin, cthr = 20,
  resl = 10000)
}
\arguments{
\item{hic_dat}{hic contact matrix for a given chromosome, either a n by n matrix, or a 3 columns data.frame
<bin1> <bin2> <counts>}

\item{hiertads_gmap}{the hierarchical domains called by GMAP}

\item{start_bin}{the start bin of the genome}

\item{end_bin}{the end bin of the genome}

\item{cthr}{the upper bound count threshold for color, default 20}

\item{resl}{reslution of Hi-C data, default 10000}
}
\description{
visualize hierarchical domains
}
