% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGMAP_source.R
\name{plotdom}
\alias{plotdom}
\title{visualize hierarchical domains}
\usage{
plotdom(hic_obj, index_obj = NULL, hiertads_gmap, chr0 = NULL,
  start_bin, end_bin, cthr = 20, resl = 10000)
}
\arguments{
\item{hic_obj}{the hic_obj in the same format as used for calling TADs}

\item{index_obj}{the index_obj used for calling TADs}

\item{hiertads_gmap}{the hierarchical domains outputted by GMAP}

\item{chr0}{the selected single chromosome ('chr1' for example)}

\item{start_bin}{the start bin of the selected chromosome}

\item{end_bin}{the end bin of the selected chromosome}

\item{cthr}{the upper bound count threshold for color, default 20}

\item{resl}{reslution of Hi-C data, default 10000, if resl=1, it works on simulated data}
}
\description{
visualize hierarchical domains
}
