% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGMAP_source.R
\name{rGMAP}
\alias{rGMAP}
\title{Detect hierarchical choromotin domains by GMAP}
\usage{
rGMAP(hic_obj, index_obj = NULL, resl = 10 * 10^3, logt = T,
  dom_order = 2, maxDistInBin = min(200, 2 * 10^6/resl), min_d = 25,
  max_d = 100, min_dp = 5, max_dp = 10, hthr = 0.95, t1thr = 0.5)
}
\arguments{
\item{hic_obj}{HiC contact matrix, generally either a 3-column data frame/data table or a tab/space delimited text file,
corresponding to the i_th, j_th bin of a chromosome and the contact frequency
\itemize{
\item For single chromosome, could also be a n by n matrix
\item For multiple chromosomes, an index_obj indicates genomic coordinate for each hic bin should be provided
}}

\item{index_obj}{A 4-columns tab/space delimited text file (or corresponding data frame/data table) indicates the genomic coordinates for each bin (compatible with HiC-Pro); with columns
\emph{bin_chr bin_start bin_end bin_id}, default NULL; when index _obj was given, it should be consitent with hic_obj}

\item{resl}{The resolution (bin size), default 10kb}

\item{logt}{Do log-transformation or not, default TRUE}

\item{dom_order}{Maximum level of hierarchical structures, default 2 (call TADs and subTADs)}

\item{maxDistInBin}{Only consider contact whose distance is not greater than maxDistInBIn bins,
default 200 bins (or 2Mb)}

\item{min_d}{The minimum d (d: window size), default 25}

\item{max_d}{The maximum d (d: window size), default 100}

\item{min_dp}{The minmum dp (dp: lower bound of tad size), defalt 5}

\item{max_dp}{The maximum dp (dp: lower bound of tad size), defalt 10.
min_d, max_d, min_dp and max_dp should be specified in number of bins}

\item{hthr}{The lower bound cutoff for posterior probability, default 0.95}

\item{t1thr}{Lower bound for t1 for calling TAD, default 0.5 quantile of test statistics
of TADs, 0.9 of subTADs}
}
\value{
A list includes following elements:
\item{tads}{A data frame with columns start, end indicates the start and end coordinates of each domain, respectively}
\item{hierTads}{A data frame with columns start, end, dom_order, where dom_order indicates the hierarchical status of a domain, 1 indicates tads, 2 indicates subtads, and so on}
\item{params}{A data frame gives the final parameters for calling TADs}
}
\description{
Detect hierarchical choromotin domains by GMAP
}
\examples{
## On simulated data:
library(rGMAP)
simu_res = data_simu('poisson-dist-hier')
true_domains = simu_res$hierTads
simu_mat = simu_res$hic_mat
predicted_domains = rGMAP(simu_mat, resl = 1)$hierTads
true_domains
predicted_domains

## On an real data example
hic_rao_IMR90_chr15   # normalized Hi-C data for IMR90, chr15 with resolution 10kb
res = rGMAP(hic_rao_IMR90_chr15, NULL, resl = 10 * 1000, dom_order = 2)
names(res)
#quickly visualize some hierarchical domains
pp = plotdom(hic_rao_IMR90_chr15, NULL, res$hierTads, NULL, 6000, 7000, 30, 10)
pp$p2
}
