% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGMAP_source.R
\name{rGMAP}
\alias{rGMAP}
\title{Detect hierarchical choromotin domains by GMAP}
\usage{
rGMAP(hic_mat, resl = 10 * 10^3, logt = T, dom_order = 2, min_d = 20,
  max_d = 100, min_dp = 5, max_dp = 10, maxDistInBin = min(200, 2 *
  10^6/resl), hthr = 0.95, t1thr = 0.5)
}
\arguments{
\item{hic_mat}{n by n matrix, n is the number of bins. Or hic_mat could by 3 columns
matrix or data.frame with columns: bin1, bin2, counts, in which bin1 and bin2, from 1 to m, are the bin
index of a hic contact}

\item{resl}{The resolution (bin size), default 10kb}

\item{logt}{Do log-transformation or not, default TRUE}

\item{dom_order}{Maximum level of hierarchical structures, default 2 (call TADs and subTADs)}

\item{min_d}{The minimum d (d: window size), default 25}

\item{max_d}{The maximum d (d: window size), default 100}

\item{min_dp}{The minmum dp (dp: lower bound of tad size), defalt 5}

\item{max_dp}{The maximum dp (dp: lower bound of tad size), defalt 10
min_d, max_d, min_dp and max_dp should be specified in number of bins}

\item{maxDistInBin}{Threshold, default 200 bins, means
only consider contacts whose distance is not greater than 200 bins}

\item{hthr}{The lower bound cutoff for posterior probability, default 0.95}

\item{t1thr}{Lower bound for t1 for calling TAD, default 0.5 quantile of test statistics
of TADs, 0.9 of subTADs}
}
\value{
A list includes following elements:
\item{tads}{A data frame with columns start, end indicates the start and end coordinates of each domain, respectively}
\item{hierTads}{A data frame with columns start, end, dom_order, where dom_order indicates the hierarchical status of a domain, 1 indicates tads, 2 indicates subtads, and so on}
\item{params}{A data frame gives the final parameters for calling TADs}
}
\description{
Detect hierarchical choromotin domains by GMAP
}
