% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGMAP1.2_source.R
\name{rGMAP}
\alias{rGMAP}
\title{Detect hierarchical choromotin domains by GMAP}
\usage{
rGMAP(hic_mat, resl = 10 * 10^3, logt = T, dom_order = 2, min_d = 50,
  Max_d = 200, min_dp = 5, Max_dp = 10, Bg_d = 200, hthr = 0.9,
  bthr = 400, t1thr = 0.75, fcthr = 0.9)
}
\arguments{
\item{hic_mat}{n by n matrix, n is the number of bins. Or hic_mat could by 3 columns
matrix or data.frame with columns: bin1, bin2, counts, in which bin1 and bin2, from 1 to m, are the bin
index of a hic contact}

\item{resl}{The resolution (bin size), default 10kb}

\item{logt}{Do log-transformation or not, default TRUE}

\item{dom_order}{Maximum level of hierarchical structures, default 2 (call TADs and subTADs)
      (if all between-within p-value < sthr (after bonferoni correction), stop call TAD or subTAD)
Max_d, Max_dp, Bg_d should be specified in number of bins and change with resloution}

\item{min_d}{The minimum d (d: window size), default 25}

\item{Max_d}{The maximum d (d: window size), default 2mb (or 200 bins)}

\item{min_dp}{The minmum dp (dp: lower bound of tad size), defalt 5}

\item{Max_dp}{The maximum dp (dp: lower bound of tad size), defalt 10}

\item{Bg_d}{The size of the background domain, dalfault 2mb or 200bins}

\item{hthr}{The lower bound cutoff for posterior probability, default 0.5 for TADs,
0.9 for subTADs}

\item{bthr}{Threshold, above which a locus' contact is ignored , default 1000, means
for each locus, only consider contacts that within 1000 bins distance away of it}

\item{t1thr}{Lower bound for t1 for calling TAD, default 0.5 quantile of test statistics
for TADs, 0.9 for subTADs}

\item{fcthr}{Lower bound quantile for fold change of proportion of 1 between
intra and inter domains, default 0.9}
}
\value{
A list includes following elements:
\item{tads}{A data frame with columns start, end indicates the start and end coordinates of each domain, respectively}
\item{hierTads}{A data frame with columns start, end, dom_order, where dom_order indicates the hierarchical status of a domain, 1 indicates tads, 2 indicates subtads, and so on}
\item{params}{A data frame gives the final parameters for calling TADs}
}
\description{
Detect hierarchical choromotin domains by GMAP
}
